fix circular imports with predicate config
fixes #3112
