+++ pyramid/config/security.py
@@ -12,9 +12,9 @@ from pyramid.interfaces import (
 
 from pyramid.csrf import LegacySessionCSRFStoragePolicy
 from pyramid.exceptions import ConfigurationError
-from pyramid.util import action_method
 from pyramid.util import as_sorted_tuple
 
+from pyramid.config.util import action_method
 
 class SecurityConfiguratorMixin(object):
