+++ pyramid/config/util.py
@@ -1,27 +1,80 @@
+import functools
 from hashlib import md5
-import inspect
+import traceback
+from zope.interface import implementer
 
 from pyramid.compat import (
     bytes_,
-    getargspec,
     is_nonstr_iter
-    )
+)
+from pyramid.interfaces import IActionInfo
 
-from pyramid.compat import im_func
 from pyramid.exceptions import ConfigurationError
+from pyramid.predicates import Notted
 from pyramid.registry import predvalseq
-
 from pyramid.util import (
     TopologicalSorter,
-    action_method,
-    ActionInfo,
-    )
+    takes_one_arg,
+)
 
-action_method = action_method # support bw compat imports
-ActionInfo = ActionInfo # support bw compat imports
+TopologicalSorter = TopologicalSorter  # support bw-compat imports
+takes_one_arg = takes_one_arg  # support bw-compat imports
+
+@implementer(IActionInfo)
+class ActionInfo(object):
+    def __init__(self, file, line, function, src):
+        self.file = file
+        self.line = line
+        self.function = function
+        self.src = src
+
+    def __str__(self):
+        srclines = self.src.split('\n')
+        src = '\n'.join('    %s' % x for x in srclines)
+        return 'Line %s of file %s:\n%s' % (self.line, self.file, src)
+
+def action_method(wrapped):
+    """ Wrapper to provide the right conflict info report data when a method
+    that calls Configurator.action calls another that does the same.  Not a
+    documented API but used by some external systems."""
+    def wrapper(self, *arg, **kw):
+        if self._ainfo is None:
+            self._ainfo = []
+        info = kw.pop('_info', None)
+        # backframes for outer decorators to actionmethods
+        backframes = kw.pop('_backframes', 0) + 2
+        if is_nonstr_iter(info) and len(info) == 4:
+            # _info permitted as extract_stack tuple
+            info = ActionInfo(*info)
+        if info is None:
+            try:
+                f = traceback.extract_stack(limit=4)
+
+                # Work around a Python 3.5 issue whereby it would insert an
+                # extra stack frame. This should no longer be necessary in
+                # Python 3.5.1
+                last_frame = ActionInfo(*f[-1])
+                if last_frame.function == 'extract_stack': # pragma: no cover
+                    f.pop()
+                info = ActionInfo(*f[-backframes])
+            except Exception: # pragma: no cover
+                info = ActionInfo(None, 0, '', '')
+        self._ainfo.append(info)
+        try:
+            result = wrapped(self, *arg, **kw)
+        finally:
+            self._ainfo.pop()
+        return result
+
+    if hasattr(wrapped, '__name__'):
+        functools.update_wrapper(wrapper, wrapped)
+    wrapper.__docobj__ = wrapped
+    return wrapper
+
 
 MAX_ORDER = 1 << 30
 DEFAULT_PHASH = md5().hexdigest()
+
 
 class not_(object):
     """
@@ -61,30 +114,6 @@ class not_(object):
     def __init__(self, value):
         self.value = value
 
-class Notted(object):
-    def __init__(self, predicate):
-        self.predicate = predicate
-
-    def _notted_text(self, val):
-        # if the underlying predicate doesnt return a value, it's not really
-        # a predicate, it's just something pretending to be a predicate,
-        # so dont update the hash
-        if val:
-            val = '!' + val
-        return val
-
-    def text(self):
-        return self._notted_text(self.predicate.text())
-
-    def phash(self):
-        return self._notted_text(self.predicate.phash())
-
-    def __call__(self, context, request):
-        result = self.predicate(context, request)
-        phash = self.phash()
-        if phash:
-            result = not result
-        return result
 
 # under = after
 # over = before
@@ -189,52 +218,3 @@ class PredicateList(object):
             score = score | bit
         order = (MAX_ORDER - score) / (len(preds) + 1)
         return order, preds, phash.hexdigest()
-
-def takes_one_arg(callee, attr=None, argname=None):
-    ismethod = False
-    if attr is None:
-        attr = '__call__'
-    if inspect.isroutine(callee):
-        fn = callee
-    elif inspect.isclass(callee):
-        try:
-            fn = callee.__init__
-        except AttributeError:
-            return False
-        ismethod = hasattr(fn, '__call__')
-    else:
-        try:
-            fn = getattr(callee, attr)
-        except AttributeError:
-            return False
-
-    try:
-        argspec = getargspec(fn)
-    except TypeError:
-        return False
-
-    args = argspec[0]
-
-    if hasattr(fn, im_func) or ismethod:
-        # it's an instance method (or unbound method on py2)
-        if not args:
-            return False
-        args = args[1:]
-
-    if not args:
-        return False
-
-    if len(args) == 1:
-        return True
-
-    if argname:
-
-        defaults = argspec[3]
-        if defaults is None:
-            defaults = ()
-
-        if args[0] == argname:
-            if len(args) - len(defaults) == 1:
-                return True
-
-    return False
